Moved string formatting functions to libmpf.py
